[ENG-9255] get proper data for Popular Pages chart rows (#732)
- Ticket: https://openscience.atlassian.net/browse/ENG-9255
- Feature flag: n/a

## Purpose

The analytics tab for registrations is meant to help users understand how viewers are using and accessing their registration. One of those metrics is popular pages. However, as you can see in the image below, there are two of the same name, which makes it confusing to understand what viewers are looking at. 

## Summary of Changes

If we are clicking over navigation in details of registration and project title saving is as expected, if go to some tab by link 
it return osf handle it to have proper title and render data in chart with row names showing what title tab was clicked